Refactor the unit tests to match the source with the notifications.
