+++ src/test/java/com/blackducksoftware/integration/hub/dataservices/notifications/transforms/PolicyOverrideTransformTest.java
@@ -1,4 +1,4 @@
-package com.blackducksoftware.integration.hub.dataservices.transforms;
+package com.blackducksoftware.integration.hub.dataservices.notifications.transforms;
 
 import static org.junit.Assert.assertEquals;
 
@@ -20,18 +20,18 @@ import com.blackducksoftware.integration.hub.api.VersionBomPolicyRestService;
 import com.blackducksoftware.integration.hub.api.component.BomComponentVersionPolicyStatus;
 import com.blackducksoftware.integration.hub.api.component.ComponentVersion;
 import com.blackducksoftware.integration.hub.api.component.ComponentVersionStatus;
-import com.blackducksoftware.integration.hub.api.notification.VulnerabilityNotificationContent;
-import com.blackducksoftware.integration.hub.api.notification.VulnerabilityNotificationItem;
-import com.blackducksoftware.integration.hub.api.notification.VulnerabilitySourceQualifiedId;
-import com.blackducksoftware.integration.hub.api.project.ProjectVersion;
+import com.blackducksoftware.integration.hub.api.notification.PolicyOverrideNotificationContent;
+import com.blackducksoftware.integration.hub.api.notification.PolicyOverrideNotificationItem;
+import com.blackducksoftware.integration.hub.api.policy.PolicyRule;
 import com.blackducksoftware.integration.hub.api.version.ReleaseItem;
 import com.blackducksoftware.integration.hub.dataservices.notifications.items.NotificationContentItem;
-import com.blackducksoftware.integration.hub.dataservices.notifications.items.VulnerabilityContentItem;
-import com.blackducksoftware.integration.hub.dataservices.notifications.transforms.VulnerabilityTransform;
+import com.blackducksoftware.integration.hub.dataservices.notifications.items.PolicyNotificationFilter;
+import com.blackducksoftware.integration.hub.dataservices.notifications.items.PolicyOverrideContentItem;
+import com.blackducksoftware.integration.hub.dataservices.notifications.transforms.PolicyViolationOverrideTransform;
 import com.blackducksoftware.integration.hub.exception.BDRestException;
 import com.blackducksoftware.integration.hub.exception.NotificationServiceException;
 
-public class VulnerabilityTransformTest {
+public class PolicyOverrideTransformTest {
 
 	private final static String PROJECT_NAME = "test project";
 	private final static String PROJECT_VERSION = "0.1.0";
@@ -39,14 +39,16 @@ public class VulnerabilityTransformTest {
 	private final static String COMPONENT_VERSION = "0.9.8";
 	private final static UUID COMPONENT_ID = UUID.randomUUID();
 	private final static UUID COMPONENT_VERSION_ID = UUID.randomUUID();
+	private final static String POLICY_NAME = "Policy Name";
+	private final static String FIRST_NAME = "myName";
+	private final static String LAST_NAME = "noMyName";
 
 	private NotificationRestService notificationService;
 	private ProjectVersionRestService projectVersionService;
 	private PolicyRestService policyService;
 	private VersionBomPolicyRestService bomVersionPolicyService;
 	private ComponentVersionRestService componentVersionService;
-	private VulnerabilityTransform transformer;
-	private List<VulnerabilitySourceQualifiedId> vulnList;
+	private PolicyViolationOverrideTransform transformer;
 
 	private NotificationRestService createNotificationService() {
 		final NotificationRestService service = Mockito.mock(NotificationRestService.class);
@@ -64,7 +66,10 @@ public class VulnerabilityTransformTest {
 	}
 
 	private PolicyRestService createPolicyService() throws IOException, BDRestException, URISyntaxException {
+		final PolicyRule rule = Mockito.mock(PolicyRule.class);
+		Mockito.when(rule.getName()).thenReturn(POLICY_NAME);
 		final PolicyRestService service = Mockito.mock(PolicyRestService.class);
+		Mockito.when(service.getPolicyRule(Mockito.anyString())).thenReturn(rule);
 		return service;
 	}
 
@@ -96,58 +101,44 @@ public class VulnerabilityTransformTest {
 		policyService = createPolicyService();
 		bomVersionPolicyService = createBomVersionService();
 		componentVersionService = createComponentVersionService();
-		vulnList = createVulnList();
-		transformer = new VulnerabilityTransform(notificationService, projectVersionService, policyService,
-				bomVersionPolicyService, componentVersionService);
+		transformer = new PolicyViolationOverrideTransform(notificationService, projectVersionService, policyService,
+				bomVersionPolicyService, componentVersionService, new PolicyNotificationFilter(null));
 
 	}
 
-	private VulnerabilityNotificationItem createNotificationItem() {
-		final VulnerabilityNotificationItem item = Mockito.mock(VulnerabilityNotificationItem.class);
-		final VulnerabilityNotificationContent content = Mockito.mock(VulnerabilityNotificationContent.class);
-		Mockito.when(item.getContent()).thenReturn(content);
+	private PolicyOverrideNotificationItem createNotificationItem() {
+		final PolicyOverrideNotificationItem item = Mockito.mock(PolicyOverrideNotificationItem.class);
+		final PolicyOverrideNotificationContent content = Mockito.mock(PolicyOverrideNotificationContent.class);
+		final List<ComponentVersionStatus> versionStatusList = new ArrayList<>();
+		final ComponentVersionStatus status = Mockito.mock(ComponentVersionStatus.class);
+		Mockito.when(content.getBomComponentVersionPolicyStatusLink()).thenReturn("PolicyRule");
 		Mockito.when(content.getComponentName()).thenReturn(COMPONENT_NAME);
-		Mockito.when(content.getVersionName()).thenReturn(COMPONENT_VERSION);
 		Mockito.when(content.getComponentVersionLink())
 		.thenReturn("/" + ComponentVersionStatus.COMPONENT_URL_IDENTIFIER + "/" + COMPONENT_ID + "/"
 				+ ComponentVersionStatus.COMPONENT_VERSION_URL_IDENTIFIER + "/" + COMPONENT_VERSION_ID);
-
-		final ProjectVersion projectVersion = Mockito.mock(ProjectVersion.class);
-		Mockito.when(projectVersion.getProjectName()).thenReturn(PROJECT_NAME);
-		Mockito.when(projectVersion.getProjectVersionName()).thenReturn(PROJECT_VERSION);
-		final List<ProjectVersion> projectVersionList = new ArrayList<>();
-		projectVersionList.add(projectVersion);
-		Mockito.when(content.getAffectedProjectVersions()).thenReturn(projectVersionList);
-		Mockito.when(content.getNewVulnerabilityCount()).thenReturn(1);
-		Mockito.when(content.getUpdatedVulnerabilityCount()).thenReturn(1);
-		Mockito.when(content.getDeletedVulnerabilityCount()).thenReturn(1);
-		Mockito.when(content.getNewVulnerabilityIds()).thenReturn(vulnList);
-		Mockito.when(content.getUpdatedVulnerabilityIds()).thenReturn(vulnList);
-		Mockito.when(content.getDeletedVulnerabilityIds()).thenReturn(vulnList);
+		versionStatusList.add(status);
+		Mockito.when(item.getContent()).thenReturn(content);
+		Mockito.when(content.getProjectName()).thenReturn(PROJECT_NAME);
+		Mockito.when(content.getFirstName()).thenReturn(FIRST_NAME);
+		Mockito.when(content.getLastName()).thenReturn(LAST_NAME);
 		return item;
-	}
-
-	public List<VulnerabilitySourceQualifiedId> createVulnList() {
-		final List<VulnerabilitySourceQualifiedId> list = new ArrayList<>();
-		final VulnerabilitySourceQualifiedId item = new VulnerabilitySourceQualifiedId("VulnSource", "vulnID");
-		list.add(item);
-		return list;
 	}
 
 	@Test
 	public void testTransform() throws Exception {
 		final List<NotificationContentItem> itemList = transformer.transform(createNotificationItem());
 		for (final NotificationContentItem item : itemList) {
-			final VulnerabilityContentItem contentItem = (VulnerabilityContentItem) item;
+			final PolicyOverrideContentItem contentItem = (PolicyOverrideContentItem) item;
 			assertEquals(PROJECT_NAME, contentItem.getProjectVersion().getProjectName());
 			assertEquals(PROJECT_VERSION, contentItem.getProjectVersion().getProjectVersionName());
 			assertEquals(COMPONENT_NAME, contentItem.getComponentName());
 			assertEquals(COMPONENT_VERSION, contentItem.getComponentVersion());
 			assertEquals(COMPONENT_ID, contentItem.getComponentId());
 			assertEquals(COMPONENT_VERSION_ID, contentItem.getComponentVersionId());
-			assertEquals(vulnList, contentItem.getAddedVulnList());
-			assertEquals(vulnList, contentItem.getUpdatedVulnList());
-			assertEquals(vulnList, contentItem.getDeletedVulnList());
+			assertEquals(FIRST_NAME, contentItem.getFirstName());
+			assertEquals(LAST_NAME, contentItem.getLastName());
+			assertEquals(1, contentItem.getPolicyRuleList().size());
+			assertEquals(POLICY_NAME, contentItem.getPolicyRuleList().get(0).getName());
 		}
 	}
 }
